Tweak/redo diameter_tcp overload protection
Commits starting at 472a080c added a throttle_cb option to diameter_tcp
to let a callback apply backpressure when it decides that additional
requests should not be read. It didn't provide a hook for knowing that
an answer was sent however, only that a spawned request process was no
longer alive. This commit adds a send callback, and redoes the existing
read/receive callback somewhat. The latter has more flexibility, and can
be used to manipulate send/recv as well as regulate the rate if incoming
traffic, while the former provides a hook for finding out when a message
has been sent.

This is still undocumented and may change. Use at your own risk.