Fix broken discard acknowledgement
A transport process can request acknowledgement of the fate of an
incoming message to a specified pid, causing it to receive one of

  {diameter, {request|answer, pid()} | discard}

depending on whether or not diameter passes the message off to a handler
process. This was broken in commit a4da06a5 (since recv/3 threw a
message that should be received), but is of little consequence since the
interface isn't yet documented and is only used from diameter_tcp with
configuration that will soon change.