Simplify acks to transport processes
Given the recent existence of hooks into both message reception and
sending, what's interesting is request that isn't answered. Acknowledge
exactly this, not the identity of handler processes. A transport process
can request acks of nonforthcoming answers by sending {diameter, {ack,
true}} to the parent peer_fsm, a handler processes identifies itself
with a {handler, pid()} message, and the peer_fsm monitors on this to be
able to send a notification to the transport if the handler dies before
sending an answer.